refactor operations apply
