+++ packages/slate/src/models/value.js
@@ -3,6 +3,7 @@ import logger from 'slate-dev-logger'
 import { Record, Set, List, Map } from 'immutable'
 
 import MODEL_TYPES from '../constants/model-types'
+import PathUtils from '../utils/path-utils'
 import Change from './change'
 import Data from './data'
 import Document from './document'
@@ -667,6 +668,474 @@ class Value extends Record(DEFAULTS) {
 
   change(attrs = {}) {
     return new Change({ ...attrs, value: this })
+  }
+  /**
+   * Add mark to text at `offset` and `length` in node by `path`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  addMark(path, offset, length, mark) {
+    let value = this
+    let { document } = value
+    document = document.addMarkByPath(path, offset, length, mark)
+    value = this.set('document', document)
+    return value
+  }
+
+  /**
+   * Insert a `node` at `path`.
+   *
+   * @param {List} path
+   * @param {Node} node
+   * @return {Value}
+   */
+
+  insertNode(path, node) {
+    let value = this
+    let { document } = value
+    document = document.insertNodeByPath(path, node)
+    value = value.set('document', document)
+
+    value = value.mapRanges(range => {
+      return range.merge({ anchorPath: null, focusPath: null })
+    })
+
+    return value
+  }
+
+  /**
+   * Insert `text` at `offset` in node by `path`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  insertText(path, offset, text, marks) {
+    let value = this
+    let { document } = value
+    let node = document.assertNode(path)
+    node = node.insertText(offset, text, marks)
+    document = document.updateNodeByPath(path, node)
+    value = value.set('document', document)
+
+    // Update any ranges that were affected.
+    value = value.clearAtomicRanges(node.key, offset)
+
+    value = value.mapRanges(range => {
+      const { anchorKey, anchorOffset, isBackward, isAtomic } = range
+
+      if (
+        anchorKey === node.key &&
+        (anchorOffset > offset ||
+          (anchorOffset === offset && (!isAtomic || !isBackward)))
+      ) {
+        return range.moveAnchor(text.length)
+      }
+
+      return range
+    })
+
+    value = value.mapRanges(range => {
+      const { focusKey, focusOffset, isBackward, isAtomic } = range
+
+      if (
+        focusKey === node.key &&
+        (focusOffset > offset ||
+          (focusOffset == offset && (!isAtomic || isBackward)))
+      ) {
+        return range.moveFocus(text.length)
+      }
+
+      return range
+    })
+
+    return value
+  }
+
+  /**
+   * Merge a node at `path` with the previous node.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  mergeNode(path) {
+    let value = this
+    const withPath = PathUtils.decrement(path)
+    let { document } = value
+    const one = document.assertNode(withPath)
+    const two = document.assertNode(path)
+    const parentPath = PathUtils.getParent(path)
+    let parent = document.getNode(parentPath)
+    const oneIndex = PathUtils.getIndex(withPath)
+    const twoIndex = PathUtils.getIndex(path)
+
+    // Perform the merge in the document.
+    parent = parent.mergeNode(oneIndex, twoIndex)
+    document = document.updateNodeByPath(parentPath, parent)
+    value = value.set('document', document)
+
+    value = value.mapRanges(range => {
+      if (two.object === 'text') {
+        const max = one.text.length
+
+        if (range.anchorKey === two.key) {
+          range = range.moveAnchorTo(one.key, max + range.anchorOffset)
+        }
+
+        if (range.focusKey === two.key) {
+          range = range.moveFocusTo(one.key, max + range.focusOffset)
+        }
+      }
+
+      range = range.merge({ anchorPath: null, focusPath: null })
+      return range
+    })
+
+    return value
+  }
+
+  /**
+   * Move a node by `path` to `newPath`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  moveNode(path, newPath) {
+    let value = this
+    const newIndex = PathUtils.getIndex(newPath)
+    let newParentPath = PathUtils.getParent(newPath)
+    const oldParentPath = PathUtils.getParent(path)
+    const oldIndex = PathUtils.getIndex(path)
+    let { document } = value
+    const node = document.assertNode(path)
+
+    // Remove the node from its current parent.
+    let parent = document.getParentByPath(path)
+    parent = parent.removeNode(oldIndex)
+    document = document.updateNodeByPath(oldParentPath, parent)
+
+    // Find the new target...
+    let target
+
+    // If the old path and the rest of the new path are the same, then the new
+    // target is the old parent.
+    if (oldParentPath.equals(newParentPath)) {
+      target = parent
+    } else if (
+      oldParentPath.every((x, i) => x === newParentPath.get(i)) &&
+      oldIndex < newParentPath.get(oldParentPath.size)
+    ) {
+      newParentPath = PathUtils.decrement(newParentPath, 1, oldParentPath.size)
+      target = document.assertNode(newParentPath)
+    } else {
+      // Otherwise, we can just grab the target normally...
+      target = document.assertNode(newParentPath)
+    }
+
+    // Insert the new node to its new parent.
+    if (!node) debugger
+    target = target.insertNode(newIndex, node)
+    document = document.updateNodeByPath(newParentPath, target)
+    value = value.set('document', document)
+
+    value = value.mapRanges(range => {
+      return range.merge({ anchorPath: null, focusPath: null })
+    })
+
+    return value
+  }
+
+  /**
+   * Remove mark from text at `offset` and `length` in node by `path`.
+   *
+   * @param {List} path
+   * @param {Number} offset
+   * @param {Number} length
+   * @param {Mark} mark
+   * @return {Value}
+   */
+
+  removeMark(path, offset, length, mark) {
+    let value = this
+    let { document } = value
+    document = document.removeMarkByPath(path, offset, length, mark)
+    value = this.set('document', document)
+    return value
+  }
+
+  /**
+   * Remove a node by `path`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  removeNode(path) {
+    let value = this
+    let { document } = value
+    const node = document.assertNode(path)
+    const first = node.object == 'text' ? node : node.getFirstText() || node
+    const last = node.object == 'text' ? node : node.getLastText() || node
+    const prev = document.getPreviousText(first.key)
+    const next = document.getNextText(last.key)
+
+    // Remove the node from the document.
+    document = document.removeNodeByPath(path)
+    value = value.set('document', document)
+
+    // Update the ranges in case they referenced the removed node.
+    value = value.mapRanges(range => {
+      const { startKey, endKey } = range
+      if (node.hasNode(startKey)) {
+        range = prev
+          ? range.moveStartTo(prev.key, prev.text.length)
+          : next ? range.moveStartTo(next.key, 0) : range.deselect()
+      }
+
+      if (node.hasNode(endKey)) {
+        range = prev
+          ? range.moveEndTo(prev.key, prev.text.length)
+          : next ? range.moveEndTo(next.key, 0) : range.deselect()
+      }
+
+      range = range.merge({ anchorPath: null, focusPath: null })
+      return range
+    })
+
+    return value
+  }
+
+  /**
+   * Remove `text` at `offset` in node by `path`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  removeText(path, offset, text) {
+    let value = this
+    const { length } = text
+    const rangeOffset = offset + length
+    let { document } = value
+
+    // Remove the text from the node and update the document.
+    let node = document.assertNode(path)
+    node = node.removeText(offset, length)
+    document = document.updateNodeByPath(path, node)
+    value = value.set('document', document)
+
+    // Update any rnages that might be affected.
+    value = value.clearAtomicRanges(node.key, offset, offset + length)
+
+    value = value.mapRanges(range => {
+      const { anchorKey } = range
+
+      if (anchorKey === node.key) {
+        return range.anchorOffset >= rangeOffset
+          ? range.moveAnchor(-length)
+          : range.anchorOffset > offset
+            ? range.moveAnchorTo(range.anchorKey, offset)
+            : range
+      }
+
+      return range
+    })
+
+    value = value.mapRanges(range => {
+      const { focusKey } = range
+
+      if (focusKey === node.key) {
+        return range.focusOffset >= rangeOffset
+          ? range.moveFocus(-length)
+          : range.focusOffset > offset
+            ? range.moveFocusTo(range.focusKey, offset)
+            : range
+      }
+
+      return range
+    })
+
+    return value
+  }
+
+  /**
+   * Set `properties` on a node by `path`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  setNode(path, properties) {
+    let value = this
+    let { document } = value
+    let node = document.assertNode(path)
+    node = node.merge(properties)
+    document = document.updateNodeByPath(path, node)
+    value = value.set('document', document)
+    return value
+  }
+
+  /**
+   * Set `properties` on mark on text at `offset` and `length` in node by `path`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  setMark(path, offset, length, mark, properties) {
+    let value = this
+    let { document } = value
+    let node = document.assertNode(path)
+    node = node.updateMark(offset, length, mark, properties)
+    document = document.updateNodeByPath(path, node)
+    value = value.set('document', document)
+    return value
+  }
+
+  /**
+   * Set `properties` on the selection.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  setSelection(properties) {
+    let value = this
+    let { document, selection } = value
+    selection = selection.merge(properties)
+    selection = selection.normalize(document)
+    value = value.set('selection', selection)
+    return value
+  }
+
+  /**
+   * Split a node by `path` at `offset`.
+   *
+   * @param {Value} value
+   * @param {Operation} operation
+   * @return {Value}
+   */
+
+  splitNode(path, position, properties) {
+    let value = this
+    let { document } = value
+
+    // Split the node by its parent.
+    const node = document.assertNode(path)
+    const index = path.last()
+    const parentPath = PathUtils.getParent(path)
+    let parent = document.getNode(parentPath)
+    parent = parent.splitNode(index, position)
+
+    // Update the properties on the new node if needed.
+    if (properties && node.object !== 'text') {
+      const p = List([index + 1])
+      let newNode = parent.getNode(p)
+      newNode = newNode.merge(properties)
+      parent = parent.updateNodeByPath(p, newNode)
+    }
+
+    // Update the node in the document.
+    document = document.updateNodeByPath(parentPath, parent)
+    value = value.set('document', document)
+
+    // Update any ranges that were affected.
+    value = value.mapRanges(range => {
+      const next = document.getNextText(node.key)
+      const { startKey, startOffset, endKey, endOffset } = range
+
+      // If the start was after the split, move it to the next node.
+      if (node.key === startKey && position <= startOffset) {
+        range = range.moveStartTo(next.key, startOffset - position)
+      }
+
+      // If the end was after the split, move it to the next node.
+      if (node.key === endKey && position <= endOffset) {
+        range = range.moveEndTo(next.key, endOffset - position)
+      }
+
+      range = range.merge({ anchorPath: null, focusPath: null })
+      return range
+    })
+
+    return value
+  }
+
+  /**
+   * Map all range objects to apply adjustments with an `iterator`.
+   *
+   * @param {Function} iterator
+   * @return {Value}
+   */
+
+  mapRanges(iterator) {
+    let value = this
+    const { document, selection, decorations } = value
+
+    if (selection) {
+      let next = selection.isSet ? iterator(selection) : selection
+      if (!next) next = selection.deselect()
+      if (next !== selection) next = next.normalize(document)
+      value = value.set('selection', next)
+    }
+
+    if (decorations) {
+      let next = decorations.map(decoration => {
+        let n = decoration.isSet ? iterator(decoration) : decoration
+        if (n && n !== decoration) n = n.normalize(document)
+        return n
+      })
+
+      next = next.filter(decoration => !!decoration)
+      next = next.size ? next : null
+      value = value.set('decorations', next)
+    }
+
+    return value
+  }
+
+  /**
+   * Remove any atomic ranges inside a `key`, `offset` and `length`.
+   *
+   * @param {String} key
+   * @param {Number} start
+   * @param {Number?} end
+   * @return {Value}
+   */
+
+  clearAtomicRanges(key, start, end = null) {
+    return this.mapRanges(range => {
+      const { isAtomic, startKey, startOffset, endKey, endOffset } = range
+      if (!isAtomic) return range
+      if (startKey !== key) return range
+
+      if (startOffset < start && (endKey !== key || endOffset > start)) {
+        return null
+      }
+
+      if (
+        end != null &&
+        startOffset < end &&
+        (endKey !== key || endOffset > end)
+      ) {
+        return null
+      }
+
+      return range
+    })
   }
 
   /**
