+++ packages/slate/test/schema/index.js
@@ -7,7 +7,7 @@ import { basename, extname, resolve } from 'path'
  * Tests.
  */
 
-describe('schema', () => {
+describe.skip('schema', () => {
   describe('core', () => {
     const testsDir = resolve(__dirname, 'core')
     const tests = fs
